Refactored all non-Windows macro constants into constexpr in their own namespaces. Also removed Windows typedefs from non-Windows classes. Random style and comment cleanups

+++ BackBeat/src/BackBeat/Audio/Synth/Synth.h
@@ -58,8 +58,8 @@ namespace BackBeat {
 	
 	private:
 		boolean m_Running;
-		UINT32 m_SamplesToRender;
-		DWORD m_SleepTime;
+		unsigned int m_SamplesToRender;
+		unsigned long m_SleepTime;
 		AudioProps m_Props;
 
 		std::shared_ptr<AudioEngine> m_Engine;
